StickerSicknessWeb/requirements.txt
Flask
python-barcode
Pillow
pylibdmtx
colorama
setuptools